- Cached services with int ids - Make selectorviewmodel usable for all, no need for inheritance and generic, control logic moved to editor modules - Jump service implementation is now much simpler - Combobox items are now populated with the user assigned names where appropriate rather than enums - Implement autofac ioc - Use Moq in tests - Split tests by dependency project - Probably a lot of things still broken, but I want to get this in

+++ RanseiLink/ViewModels/ModelViewModels/MapGridCellViewModel.cs
@@ -43,7 +43,7 @@ public class MapGridCellViewModel : ViewModelBase
             if (Gimmicks.Any())
             {
                 var gimmick = Gimmicks.Last();
-                GimmickImagePath = _spriteProvider.GetSpriteFile(SpriteType.StlStageObje, _gimmickService.Retrieve(gimmick.Gimmick).Image).File;
+                GimmickImagePath = _spriteProvider.GetSpriteFile(SpriteType.StlStageObje, _gimmickService.Retrieve((int)gimmick.Gimmick).Image).File;
             }
             else
             {
